\name{wardrobe-package}
\alias{wardrobe-package}
\docType{package}
\title{
Wardrobe package
}
\description{
Wardrobe package provides tools to access and process Wardrobe datasets
}
\details{
\tabular{ll}{
Package: \tab wardrobe\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2015-02-17\cr
License: \tab MIT\cr
URL:  \tab \url{https://wardrobe.porter.st/}\cr
}
}
\author{
Andrey Kartashov
 \email{porter@porter.st}
}
\references{
%~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
  \code{\link{wardrobe}}
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
library("wardrobe")
experiment <- wardrobe(id=1)
}
