\name{wardrobe-package}
\alias{wardrobe-package}
\alias{wardrobe}
\docType{package}
\title{
Wardrobe package
}
\description{
Wardrobe package provides tools to access Wardrobe datasets
}
\details{
\tabular{ll}{
Package: \tab wardrobe\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-17\cr
License: \tab MIT\cr
}
experiment <- wardrobe(id)
}
\author{
Andrey Kartashov
Maintainer: Andrey Kartashov <porter@porter.st>
}
\references{
%~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
experiment <- wardrobe(id=1)
}
